requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
sentence-transformers
faiss-cpu
